requirements.txt
fastapi
uvicorn
openai
python-dotenv
web3
slither-analyzer